Delete attempt of 16-bit preciscion

+++ BaseModel/SSD/ssd/modeling/box_head/loss.py
@@ -26,7 +26,7 @@ class MultiBoxLoss(nn.Module):
         num_classes = confidence.size(2)
         with torch.no_grad():
             # derived from cross_entropy=sum(log(p))
-            loss = -F.log_softmax(confidence, dim=2)[:, :, 0].half()
+            loss = -F.log_softmax(confidence, dim=2)[:, :, 0]
             mask = box_utils.hard_negative_mining(loss, labels, self.neg_pos_ratio)
 
         confidence = confidence[mask, :]
